+++ src/dcegm/solve.py
@@ -6,8 +6,10 @@ from typing import Tuple
 import numpy as np
 import pandas as pd
 from dcegm.egm import compute_optimal_policy_and_value
-from dcegm.egm import get_child_state_marginal_util_and_exp_max_value
 from dcegm.integration import quadrature_legendre
+from dcegm.marg_utilities_and_exp_value import (
+    get_child_state_marginal_util_and_exp_max_value,
+)
 from dcegm.pre_processing import create_multi_dim_arrays
 from dcegm.pre_processing import get_partial_functions
 from dcegm.state_space import get_child_indexes
