+++ tests/test_egm.py
@@ -3,12 +3,12 @@ from itertools import product
 
 import numpy as np
 import pytest
-from dcegm.egm import calc_choice_probability
-from dcegm.egm import get_child_state_choice_specific_policy
-from dcegm.egm import get_child_state_choice_specific_values
-from dcegm.egm import get_child_state_marginal_util
 from dcegm.interpolate import interpolate_policy
 from dcegm.interpolate import interpolate_value
+from dcegm.marg_utilities_and_exp_value import calc_choice_probability
+from dcegm.marg_utilities_and_exp_value import get_child_state_choice_specific_policy
+from dcegm.marg_utilities_and_exp_value import get_child_state_choice_specific_values
+from dcegm.marg_utilities_and_exp_value import get_child_state_marginal_util
 from dcegm.pre_processing import calc_current_value
 from numpy.testing import assert_array_almost_equal as aaae
 from scipy.special import roots_sh_legendre
